aeo-checks/master_requirements.txt
# Master AEO Service Requirements
# Combines dependencies from both aeo-health-check and aeo-leaderboard

# Core Framework
fastapi>=0.104.1
uvicorn[standard]>=0.24.0
pydantic>=2.5.0

# HTTP and Async
httpx>=0.25.2
aiofiles>=23.2.1

# HTML/Web Processing  
beautifulsoup4>=4.12.2
lxml>=4.9.3
playwright>=1.40.0

# AI and LLM Integration
openai>=1.3.0
anthropic>=0.7.0
google-generativeai>=0.3.0

# Data Processing
pandas>=2.1.4
numpy>=1.25.2

# Utilities
python-dateutil>=2.8.2
python-multipart>=0.0.6
python-dotenv>=1.0.0

# Logging and Monitoring
structlog>=23.2.0
psutil>=5.9.6

# Additional for Master Service
difflib  # Built-in for brand confusion analysis
pathlib  # Built-in for file handling
time     # Built-in for performance tracking
asyncio  # Built-in for parallel processing
re       # Built-in for pattern matching
logging  # Built-in for enhanced logging